add statsmodel

+++ gwas/aou/aou_phewas.py
@@ -117,7 +117,7 @@ def main():
         #else:
             #add linear regresssion
        #model = OLS(ptdata["phenotype"], ptdata[["genotype"]+covars])
-        model = sm.OLS(ptdata["phenotype"], ptdata[["genotype"]+covars])
+        model = sm.OLS(ptdata["phenotype"], ptdata[["genotype"].astype(float)+covars])
     	# is case/control or quantitative. if case/control, use
     	# logistic regression instead
